Refactored the theme to use ts
